log,swap: extracted swap log into swap package
